pyproject.toml
[tool.poetry]
name = "amlhpc"
version = "0.1.9"
authors = [ "Hugo Meiland <hugo.meiland@microsoft.com>", "Utkarsh Ayachit <uayachit@microsoft.com>" ]
description = "Emulate Slurm/PBS/LSF HPC scheduler in Azure ML"
readme = "README.md"
packages = [
    { include = "amlhpc" }
]

[tool.poetry.scripts]
sbatch = 'amlhpc.slurm.sbatch:sbatch'
sinfo = 'amlhpc.slurm.sinfo:sinfo'
squeue = 'amlhpc.slurm.squeue:squeue'

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
azure-ai-ml = "^1.12.1"
#azureml-core = "^1.54.0.post1"
#azure-cli = "^2.49.0"

[project.urls]
Homepage = "https://github.com/hmeiland/aml-slurm"
Issues = "https://github.com/hmeiland/aml-slurm/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
